+++ Projects/UserInterface/Localization/Resources/en.lproj/Global.strings
@@ -1,0 +1,9 @@
+/* 
+  Global.strings
+  Localization
+
+  Created by 최형우 on 2023/06/07.
+  Copyright © 2023 com.msg. All rights reserved.
+*/
+
+"unknown_error" = "An unknown error occurred, please contact us if it persists.";
