+++ Projects/UserInterface/Localization/Resources/en.lproj/Signin.strings
@@ -13,3 +13,4 @@
 "signup_title" = "Signup";
 "signup_button_title" = "New to Dotori? Signup";
 "find_password_button_title" = "Find Password";
+"invalid_password_error" = "Email or password doesn't match.";
